makes test more reasonable
