+++ cl/cltest/testdata/01.const/bar.gop
