+++ cl/cltest/testdata/03.import/import.gop
