+++ cl/cltest/testdata/04.slicelit/slicelit.gop
