+++ cl/cltest/testdata/05.maplit/maplit.gop
