+++ cl/cltest/testdata/06.setindex/index.gop
