+++ cl/cltest/testdata/07.string-map-slice/datastruct.gop
