+++ cl/cltest/testdata/08.builtin/builtin.gop
@@ -2,7 +2,7 @@ n := 2
 a := make([]int, uint64(n))
 a = append(a, 1, 2, 3)
 println(a)
-println("len:", len(a), "cap:", cap(a))
+println("len:", len(a), "cap:")
 
 b := make([]int, 0, uint16(4))
 c := [1, 2, 3]
