+++ cl/cltest/testdata/09.list-comprehens/list_comprehens.gop
