+++ cl/cltest/testdata/10.map-comprehens/map_comprehens.gop
