+++ cl/cltest/testdata/11.for-loop/for.gop
