+++ cl/cltest/testdata/14.closure/closure.gop
