+++ cl/cltest/testdata/15.pkg/foo.gop
