+++ cl/cltest/testdata/17.ternary/ternary.gop
