+++ cl/cltest/testdata/18.fib/fib.gop
